+++ QuailServer/Main/main.swift
@@ -1,0 +1,4 @@
+import Foundation
+import QuailServerCore
+
+await exit(QuailServerApp.run(arguments: Array(CommandLine.arguments.dropFirst())))